% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{aml_store_push}
\alias{aml_store_push}
\title{Push local targets metadata back to the cloud store}
\usage{
aml_store_push(cloud_store_base, project = Sys.getenv("TAR_PROJECT"))
}
\arguments{
\item{cloud_store_base}{Character scalar. Parent directory of the targets
store on Azure Blob storage.}

\item{project}{Character scalar. Target project/region ID. Defaults to the
`TAR_PROJECT` environment variable.}
}
\value{
Invisibly, `TRUE` on success.
}
\description{
Push local targets metadata back to the cloud store
}
