% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{aml_store_reset}
\alias{aml_store_reset}
\title{Reset the local targets cache and parity symlinks}
\usage{
aml_store_reset(project = Sys.getenv("TAR_PROJECT"), link_dir = "data")
}
\arguments{
\item{project}{Character scalar. Target project/region ID. Defaults to the
`TAR_PROJECT` environment variable.}

\item{link_dir}{Character scalar. Parent directory containing parity
symlinks. Defaults to `"data"`.}
}
\value{
Invisibly, `TRUE`.
}
\description{
Reset the local targets cache and parity symlinks
}
