% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{read_token}
\alias{read_token}
\title{Read the Run ID from a token file}
\usage{
read_token(target_name)
}
\arguments{
\item{target_name}{Character. Target name.}
}
\value{
Run ID (character scalar).
}
\description{
Read the Run ID from a token file
}
\keyword{internal}
