% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{token_path}
\alias{token_path}
\title{Path to the JSON token file for a target}
\usage{
token_path(target_name)
}
\arguments{
\item{target_name}{Character. Target name.}
}
\value{
Character file path inside the active targets store.
}
\description{
Path to the JSON token file for a target
}
\keyword{internal}
