% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens.R
\name{write_token}
\alias{write_token}
\title{Write a job token file}
\usage{
write_token(target_name, run_id)
}
\arguments{
\item{target_name}{Character. Target name.}

\item{run_id}{Character. Azure ML Run ID.}
}
\value{
Invisibly, the path written.
}
\description{
Write a job token file
}
\keyword{internal}
