% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amltargets-package.R
\docType{package}
\name{amltargets-package}
\alias{amltargets}
\alias{amltargets-package}
\title{amltargets: Azure ML Target Factory for the targets Package}
\description{
Provides a targets Target Factory (Archetype) that offloads heavy targets to Azure ML Compute Clusters as asynchronous Command Jobs. Job state is tracked via JSON token files on a shared Azure Blob storage mount so that pipelines survive local computer sleep and reboot cycles.
}
\author{
\strong{Maintainer}: Jeffery Leirness \email{jeffery.leirness@noaa.gov} (\href{https://orcid.org/0000-0003-3604-2790}{ORCID})

}
\keyword{internal}
