% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-cli.R
\name{delete_token}
\alias{delete_token}
\title{Delete a token file}
\usage{
delete_token(target_name)
}
\arguments{
\item{target_name}{Character. Target name.}
}
\value{
Invisibly, the result of `unlink()`.
}
\description{
Delete a token file
}
\keyword{internal}
