% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-cli.R
\name{token_exists}
\alias{token_exists}
\title{Check whether a token file exists}
\usage{
token_exists(target_name)
}
\arguments{
\item{target_name}{Character. Target name.}
}
\value{
Logical scalar.
}
\description{
Check whether a token file exists
}
\keyword{internal}
